Moved utils to lib
