+++ lib/mysql.js
@@ -6,7 +6,7 @@ const logger = require('@janiscommerce/logger');
 const QueryBuilder = require('@janiscommerce/query-builder');
 
 const MySQLError = require('./mysql-error');
-const Utils = require('../utils');
+const Utils = require('./utils');
 
 const CONNECTION_LIMIT = 10;
 const DEFAULT_LIMIT = 500;
